+++ cli/types.ts
@@ -33,10 +33,14 @@ export type Config = {
 };
 
 export type CanisterConfig = {
-  main: string;
+  main?: string;
   args?: string[];
   candid?: string;
   initArg?: string;
+  "check-stable"?: {
+    path: string;
+    skipIfMissing?: boolean;
+  };
 };
 
 export type Dependencies = Record<string, Dependency>;